Remove old cli
