+++ crates/edit_prediction_cli/Cargo.toml
@@ -9,7 +9,7 @@ license = "GPL-3.0-or-later"
 workspace = true
 
 [[bin]]
-name = "ep_cli"
+name = "ep"
 path = "src/main.rs"
 
 [dependencies]
@@ -23,6 +23,7 @@ cloud_llm_client.workspace= true
 cloud_zeta2_prompt.workspace = true
 collections.workspace = true
 debug_adapter_extension.workspace = true
+dirs.workspace = true
 edit_prediction_context.workspace = true
 extension.workspace = true
 fs.workspace = true
@@ -55,6 +56,7 @@ toml.workspace = true
 util.workspace = true
 watch.workspace = true
 edit_prediction = { workspace = true, features = ["eval-support"] }
+wasmtime.workspace = true
 zlog.workspace = true
 
 [dev-dependencies]
