+++ crates/edit_prediction_cli/src/format_prompt.rs
