+++ crates/edit_prediction_cli/src/score.rs
