+++ crates/edit_prediction_cli/src/teacher.prompt.md
