--- crates/edit_prediction_cli/src/training/mod.rs
@@ -1,4 +1,0 @@
-pub mod context;
-pub mod distill;
-pub mod llm_client;
-pub mod teacher;
