--- crates/edit_prediction_cli2/Cargo.toml
@@ -1,66 +1,0 @@
-[package]
-name = "edit_prediction_cli2"
-version = "0.1.0"
-edition.workspace = true
-publish.workspace = true
-license = "GPL-3.0-or-later"
-
-[lints]
-workspace = true
-
-[[bin]]
-name = "ep"
-path = "src/main.rs"
-
-[dependencies]
-anyhow.workspace = true
-anthropic.workspace = true
-http_client.workspace = true
-chrono.workspace = true
-clap.workspace = true
-client.workspace = true
-cloud_llm_client.workspace= true
-cloud_zeta2_prompt.workspace = true
-collections.workspace = true
-debug_adapter_extension.workspace = true
-dirs.workspace = true
-edit_prediction_context.workspace = true
-extension.workspace = true
-fs.workspace = true
-futures.workspace = true
-gpui.workspace = true
-gpui_tokio.workspace = true
-indoc.workspace = true
-language.workspace = true
-language_extension.workspace = true
-language_model.workspace = true
-language_models.workspace = true
-languages = { workspace = true, features = ["load-grammars"] }
-log.workspace = true
-node_runtime.workspace = true
-paths.workspace = true
-project.workspace = true
-prompt_store.workspace = true
-pulldown-cmark.workspace = true
-release_channel.workspace = true
-reqwest_client.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-settings.workspace = true
-shellexpand.workspace = true
-smol.workspace = true
-sqlez.workspace = true
-sqlez_macros.workspace = true
-terminal_view.workspace = true
-toml.workspace = true
-util.workspace = true
-watch.workspace = true
-edit_prediction = { workspace = true, features = ["eval-support"] }
-wasmtime.workspace = true
-zlog.workspace = true
-
-[dev-dependencies]
-indoc.workspace = true
-gpui = { workspace = true, features = ["test-support"] }
-project = { workspace = true, features = ["test-support"] }
-pretty_assertions.workspace = true
