--- crates/edit_prediction_cli2/LICENSE-GPL
@@ -1,1 +1,0 @@
-../../LICENSE-GPL
